Move call_with_backoff tests to test_messaging.py
The helper moved to messaging.operations; its tests follow. Renamed
TestRateLimitHandling -> TestCallWithBackoff to match the file's
function-under-test naming; imports hoisted to module level.
